fail on failure

+++ build_ltsmin.sh
@@ -195,9 +195,11 @@ pushd $ROOTDIR
 
 pwd
 ls 
-cd lts_install_dir/bin ; mkdir pp ; \rm *-dist *-sym pins2torx ; mv pins2* pp/ ; mv ltl2* pp/ ; \rm * ; mv pp/* . ; rm -rf pp/ ; for i in * ; do strip $i ; done; cd ../..
-tar czvf ltsmin_linux_64.tar.gz lts_install_dir/
-cp ltsmin_linux_64.tar.gz website/
-ls
-ls -lah website/
+if [ -f lts_install_dir/bin/pins2lts-seq ]; then
+	cd lts_install_dir/bin ; mkdir pp ; \rm *-dist *-sym pins2torx ; mv pins2* pp/ ; mv ltl2* pp/ ; \rm * ; mv pp/* . ; rm -rf pp/ ; for i in * ; do strip $i ; done; cd ../..
+	tar czvf ltsmin_linux_64.tar.gz lts_install_dir/
+	cp ltsmin_linux_64.tar.gz website/
+else 
+	exit 1;
+fi
